+++ zarrs/src/array/codec/array_to_array/fixedscaleoffset/fixedscaleoffset_codec.rs
@@ -67,6 +67,7 @@ macro_rules! unsupported_dtypes {
                 unit: _,
                 scale_factor: _,
             }
+            | DataType::Optional(_)
             | DataType::Extension(_)
     };
 }