+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `VariableLengthBytes`
 - Add `ArrayBytesDecodeIntoTarget`
+- Add `OptionalCodec`
+- Add `DataType::Optional`
+- Implement `Element` for `Option<T>` where `T: Element`
+- Implement `ElementOwned` for `Option<T>` where `T: ElementOwned`
+- Implement `ElementFixedLength` for `Option<T>` where `T: ElementFixedLength`
 
 ### Changed
 
@@ -24,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `ArrayToBytesCodecTraits::decode_into()`
   - `[Async]Array::retrieve_chunk[_subset]_into()`
 - **Breaking**: `ArrayBytes::validate()` now takes a `DataType` instead of a `DataTypeSize`
+- **Breaking**: Mark `ArrayBytes` as non-exhaustive
 
 ### Removed